+++ internal/coreinternal/processor/filterlog/filterlog_test.go
@@ -40,14 +40,15 @@ func TestLogRecord_validateMatchesConfiguration_InvalidConfig(t *testing.T) {
 		{
 			name:        "empty_property",
 			property:    filterconfig.MatchProperties{},
-			errorString: `at least one of "attributes", "libraries", "resources" or "log_bodies" field must be specified`,
+			errorString: `at least one of "attributes", "libraries", "resources", "log_bodies" or "log_severities" field must be specified`,
 		},
 		{
 			name: "empty_log_bodies_and_attributes",
 			property: filterconfig.MatchProperties{
 				LogBodies: []string{},
+				LogSeverities: []string{},
 			},
-			errorString: `at least one of "attributes", "libraries", "resources" or "log_bodies" field must be specified`,
+			errorString: `at least one of "attributes", "libraries", "resources", "log_bodies" or "log_severities" field must be specified`,
 		},
 		{
 			name: "span_properties",
@@ -156,11 +157,19 @@ func TestLogRecord_Matching_True(t *testing.T) {
 				LogBodies: []string{"AUTH.*"},
 			},
 		},
+		{
+			name: "log_severity_regexp_match",
+			properties: &filterconfig.MatchProperties{
+				Config:        *createConfig(filterset.Regexp),
+				LogSeverities: []string{"debug.*"},
+			},
+		},
 	}
 
 	lr := pdata.NewLogRecord()
 	lr.Attributes().InsertString("abc", "def")
 	lr.Body().SetStringVal("AUTHENTICATION FAILED")
+	lr.SetSeverityText("debug")
 
 	for _, tc := range testcases {
 		t.Run(tc.name, func(t *testing.T) {