+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - `datadogexporter`: Add `host_metadata` configuration section to configure host metadata export (#9100)
 
 - `cmd/mdatagen`: Update documentation generated for attributes to list enumerated values and show the "value" that will be visible on metrics when it is different from the attribute key in metadata.yaml (#8983)
+- `attributesprocessor`: Support filter by severity (#9132)
 
 ### 🛑 Breaking changes 🛑